If we get 3 consecutive timeout when we read char on the serial port, we close and reopen the port.
Use of a flag in order to state the serial port is not ready. It gets
ready again once we can read the firmware code.
So, the connector is ready once the port is opened and the flag at true.

We attempt 3 times to get the firmware code before stating the port is
not ready.

We purge the RX and TX buffer before sending a command.

We shutdown the executor when we interrupt the Virtuino conector (it was
a bug).

The function isConnectorReady is checked in each VirtuinoConnectorr
public functions.

The managesubscription check the isConnectorReady only once at the
beginning (not at each call of getFloat).